[MODIFY] makefile_completer.py Separing the directory finder and the defines
