+++ scripts/definers.py
@@ -1,0 +1,3 @@
+IDENTIFIERS = ["src", "tests", "my"]
+PATHS = ["\t$(SRC_DIR)/", "\t$(TEST_SRC)/", "\t"]
+MAKEFILE_VAR = ["SRC\t=", "TESTS\t=", "SRC\t="]
